Add files via upload
In this project, I used the API service from a cryptocurrency website, CoinMarketCap. I created a script to extract the data and append a few pulls. The end of the script, there are a few basic lines of code that would show how to visualize some of the data I pulled from.